Transfer canvasRef, handleDrawingHangman from the HangmanCanvas component to the GameContext for more access
