+++ src/context/GameContextProvider.jsx
@@ -1,4 +1,5 @@
-import React, { useState, createContext, useContext } from 'react';
+import React, { useState, createContext, useContext, useRef } from 'react';
+import drawOnCanvas from '../helpers/drawingOnCanvasHelpers';
 
 export const GameContext = createContext();
 
@@ -9,9 +10,20 @@ export const getContextValues = () => {
 const GameContextProvider = ({ children }) => {
   const [wrongGuessCount, setWrongGuessCount] = useState(0);
 
+  const canvasRef = useRef(null);
+
+  const handleDrawingHangman = (canvas, count) => {
+    if (count < 10) {
+      drawOnCanvas(canvas, count);
+      setWrongGuessCount((prev) => prev + 1);
+    }
+  };
+
   const contextValue = {
     wrongGuessCount,
     setWrongGuessCount,
+    canvasRef,
+    handleDrawingHangman,
   };
 
   return (
